fix: renamed helperFuncs.js to helpers.js
